WA-230 Refactoring safe.builder to provide only safe records
