+++ src/routes/safeList/components/Layout.stories.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react'
 import { List } from 'immutable'
 import * as React from 'react'
 import styles from '~/components/layout/PageFrame/index.scss'
-import { SafeFactory } from '~/routes/safe/store/test/builder/index.builder'
+import { SafeFactory } from '~/routes/safe/store/test/builder/safe.builder'
 import Component from './Layout'
