--- src/routes/safeList/store/test/safe.spec.js
@@ -1,7 +1,0 @@
-// @flow
-import safesSelectorTests from './safes.selector'
-
-describe('SafeList Test suite', () => {
-  // safesSelector SELECTOR
-  safesSelectorTests()
-})
